style: organiza arquivos
