+++ backend/main.py
@@ -1,7 +1,8 @@
 from fastapi import FastAPI
-from routes import data
+from routes import documents, experimentos
 
 app = FastAPI()
 
 
-app.include_router(data.router, prefix='/data', tags=['data'])
+app.include_router(experimentos.router, prefix='/experimentos', tags=['experimentos'])
+app.include_router(documents.router, prefix='/documents', tags=['documents'])
